Moved component description model to duit_core package
